♻️ refactor: dto changes (from record to class) and added the modelma…

+++ src/main/java/br/edu/ifpb/easyschoolback/business/service/GradeService.java
@@ -1,9 +1,6 @@
 package br.edu.ifpb.easyschoolback.business.service;
 
 
-import br.edu.ifpb.easyschoolback.business.mappers.GradeRequestToGradeMapper;
-import br.edu.ifpb.easyschoolback.business.mappers.GradeToGradeResponseMapper;
-import br.edu.ifpb.easyschoolback.business.mappers.StudentRequestToStudentMapper;
 import br.edu.ifpb.easyschoolback.model.entities.Grade;
 import br.edu.ifpb.easyschoolback.model.entities.Student;
 import br.edu.ifpb.easyschoolback.model.repository.GradeRepository;
@@ -11,7 +8,7 @@ import br.edu.ifpb.easyschoolback.model.repository.StudentRepository;
 import br.edu.ifpb.easyschoolback.model.repository.exception.EntityNotFoundException;
 import br.edu.ifpb.easyschoolback.presentation.dtos.grade.CreateGradeRequestDto;
 import br.edu.ifpb.easyschoolback.presentation.dtos.grade.GradeResponseDto;
-import br.edu.ifpb.easyschoolback.presentation.dtos.student.StudentResponseDto;
+import br.edu.ifpb.easyschoolback.presentation.dtos.grade.UpdateGradeRequestDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
@@ -24,12 +21,12 @@ public class GradeService {
 
     private final GradeRepository gradeRepository;
     private final StudentRepository studentRepository;
-
+    private final ModelMapper mapper;
 
     public GradeResponseDto assignGradeToStudent(final CreateGradeRequestDto gradeRequest, final Long studentId) {
         log.info("Assigning grade to student: {}", studentId);
 
-        Grade grade = GradeRequestToGradeMapper.mapper(gradeRequest);
+        Grade grade = mapper.map(gradeRequest, Grade.class);
 
         Student student = studentRepository.findById(studentId).orElseThrow(
                 EntityNotFoundException::new);
@@ -37,18 +34,39 @@ public class GradeService {
         grade.setStudent(student);
 
         gradeRepository.save(grade);
-        return GradeToGradeResponseMapper.mapper(grade);
+        return mapper.map(grade, GradeResponseDto.class);
     }
 
-    public void removeGradeFromStudent(final Long gradeId) {
+    public void removeGradeFromStudent(final Long gradeId, final Long studentId) {
         log.info("Removing grade from student: {}", gradeId);
 
         Grade grade = findGradeEntityById(gradeId);
+        Student student = studentRepository.findById(studentId).orElseThrow(EntityNotFoundException::new);
+
+        if (!grade.getStudent().equals(student)) {
+            throw new EntityNotFoundException();
+        }
 
         gradeRepository.delete(grade);
+
+        log.info("Grade removed from student: {}", gradeId);
     }
 
+    public GradeResponseDto updateGrade(final Long gradeId, final UpdateGradeRequestDto updateRequest) {
 
+        log.info("Updating grade: {}", updateRequest);
+
+        Grade grade = findGradeEntityById(gradeId);
+
+        grade.setGradeValue(updateRequest.getGradeValue());
+        grade.setWeight(updateRequest.getWeight());
+        grade.setType(updateRequest.getType());
+
+        gradeRepository.save(grade);
+
+        log.info("Grade updated: {}", gradeId);
+        return mapper.map(grade, GradeResponseDto.class);
+    }
 
     private Grade findGradeEntityById(final Long gradeId) {
         return gradeRepository.findById(gradeId).orElseThrow(
